requirements.txt
asyncpg
bcrypt
fastapi
jwt
passlib
python-dotenv
uvicorn[standard]
